tests/requirements-test.txt
# Testing dependencies for Live Graph System

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0
pytest-mock>=3.10.0

# Coverage reporting
coverage>=7.0.0
coverage[toml]>=7.0.0

# Performance and memory testing
psutil>=5.9.0
memory-profiler>=0.60.0

# HTTP testing
requests-mock>=1.10.0
responses>=0.22.0

# Web testing (for Puppeteer integration)
selenium>=4.8.0
webdriver-manager>=3.8.0

# Data validation and testing
jsonschema>=4.17.0
faker>=18.0.0

# Async testing
pytest-asyncio>=0.21.0

# Database testing (if needed)
pytest-postgresql>=4.1.0
pytest-redis>=3.0.0

# Mocking and fixtures
factory-boy>=3.2.0
freezegun>=1.2.0

# Linting and code quality (for CI/CD)
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Documentation testing
pytest-doctestplus>=0.12.0

# Parallel testing
pytest-parallel>=0.1.0

# Test reporting
pytest-json-report>=1.5.0
pytest-benchmark>=4.0.0

# Security testing
bandit>=1.7.0
safety>=2.3.0

# Load testing
locust>=2.14.0

# API testing
tavern>=1.23.0

# Browser automation (for E2E tests)
playwright>=1.30.0

# Test data generation
hypothesis>=6.68.0

# XML/HTML parsing for test validation
lxml>=4.9.0
beautifulsoup4>=4.11.0

# Time and date testing
freezegun>=1.2.0
time-machine>=2.9.0

# Network testing
pyfakefs>=5.1.0
responses>=0.22.0

# Configuration testing
pytest-env>=0.8.0
python-dotenv>=1.0.0
